+++ cursepy/__init__.py
@@ -9,5 +9,5 @@ from cursepy.wrapper import CurseClient
 
 # Define some metadata here:
 
-__version__ = '1.1.3'
+__version__ = '1.2.0'
 __author__ = 'Owen Cochell'